make graph optional dependency
